Refactor main logic into separate modules
